+++ CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Bug fixes
 
+# 2.5.19 (5 Feb 2026)
+
+- `DetailedTrace`: add ActiveRecord backend, generator #5525
+- `DetailedTrace`: improve ActiveRecord backend, filter hashes before logging #5527
+
+# 2.5.18 (22 Jan 2026)
+
+- `GraphQL::Dashboard`: properly require `action_controller` before using it #5510
+- `GraphQL::Dashboard`: don't use `config.asset_host` for Dashboard assets since they're handled by the dashboard itself #5511
+
 # 2.5.17 (21 Jan 2026)
 
 ### Bug fixes